bot/requirements.txt
nextcord 
python-dotenv